minor changes in operations & main file

+++ src/sumitvar/evaluator/main/Evaluator.java
@@ -6,6 +6,6 @@ public class Evaluator {
     public static void main(String[] args) {
         EvaluatorOperations operation = new EvaluatorOperations();
         float result =  operation.executeEvaluator(args);
-        System.out.println(result);
+        System.out.println((int)result);
     }
 }
